Updated README
There are a few things in here that still need to be updated as we generate them. For example, the THIRD_PARTY_NOTICES.md that is referenced under the License does not exist yet, nor does the nvmrc. Additionally, there is no support thread for this project yet (and I'm not sure if we want to generate one just yet).