Made changes related to navigation pane editing
I made these changes because updating the navigation pane is going to be a regular task for tech writers, and if it's hidden over in Contributing.md, it would be difficult to find.

In Contributing.md, I replaced the content about the navigation pane with a link to our style guide.

I created a new topic in our style guide called "update-left-navigation-pane.mdx" and inserted the original content (with a few edits) from the .md file.