Added schema definition for `grid_column_count`
### Summary

Per [this thread](https://discuss.newrelic.com/t/100354) on the Explorers Hub community forum, a new field—`grid_column_count`—was added to the dashboard's data definition schema. For ease of consistent reference, this implementation should be reflected in the official documentation as well.

### Changes

In the following places:

* In the "Example dashboard schema" code snippet, added `"grid_column_count": 3|12`.
* In the "Dashboard data definitions" table, added data element entry for `grid_column_count`.

### To-do

Follow-up checks for the reviewer:

* Confirm whether type is actually `Integer` or some other numeric type like `Long`.
* Confirm whether values other than 3 and 12 are supported in any capacity.
* Amend the "Important" and "Caution" call-outs regarding dashboard's widget size limits and column configuration, in accordance with `grid_column_count`.


//
Thank you.