Fix PageView and PageViewTiming links
When accessing the links on the [Events reported by Browser Monitoring page](https://docs.newrelic.com/docs/insights/event-data-sources/default-events/events-reported-browser-monitoring/), several of the links only go to the `AjaxRequest` event and not the correct event.

I could not find the proper `event_tids` to use, so I used the resolved URL for the specific event pages. I have only fixed the `PageView` and `PageViewTiming` events. This may not work for you all, but I thought I'd throw this out there.