Updated vector config samples
Updated the sample config file to include a line for encoding. That's required to get vector to work. I also deleted the data_dir lines from the config since that entry caused an error with vector on startup. Removing that line fixes the issue. It's also redundant since that data_dir entry comes with the default vector.toml file. Last, I updated the vector.conf file to accurately reflect the file having a .toml extension, not a .conf extension.