mysql and postgres doc update
Added an entry for mysql installation on Windows. NR hosts the windows .msi file on our download site but we don't have directions for it in our docs. The docs explicity required a linux host which isn't true so I cleaned up that verbiage as well. For both files i just updated our links to the download site to default to https since we don't redirect there from http links.