Clarify violation close for LoS violations
As an engineer working in the platform, I was unsure what conditions needed to be met for a loss of signal violation to automatically close. Does the signal need to resume and maintain across the chosen expiration window? Does it immediately resume once it's detected?  I tested this with a synthetics monitor and the violation closed immediately once a new data point was detected. Hopefully this can help  our customers who have similar questions.