Clarify limits of ignoring through configuration
The configuration method `ignore_url_regexes` only prevents Transaction events reporting and not spans/traces. Related to:

* https://newrelic.atlassian.net/browse/GTSE-10084

Note: there is no engineering response in the above Jira as we discussed this in GTS pairing and determined this is how the configuration method is expected to work. Feel free to ping #ruby-agent or #gts-ruby for any questions!

Since new releases of the Ruby agent will overwrite any edits to the docs site [configuration page](https://docs.newrelic.com/docs/agents/ruby-agent/configuration/ruby-agent-configuration/#rules-ignore_url_regexes), I submitted a Github issue to the Ruby team to specify this in the configuration doc:

https://github.com/newrelic/newrelic-ruby-agent/issues/729