chore: remove use of ERRORED state
* we currently dont have a way to change the state of errored
translations so rather than use it (and have to manually change the
state ourselves), we are removing it for the moment. that means any
translation that would have previously been marked as errored is now
left in a PENDING state, which will be picked up again on the next run.