Update Ruby ∞T requirements
Current phrasing caused confusion on whether or not the Ruby agent also requires the separate Infinite Tracing gem to be installed. I attempted to clarify that it is required by tacking on the Infinite Tracing gem in the 'Install or update' step but I may have just created a run-on sentence so re-phrase as needed!

I also bumped the min. required version to v7 so new customers do not encounter this bug:
https://github.com/newrelic/newrelic-ruby-agent/pull/629