fix: adding yarn lock back
It looks like the lockfile was removed in the following commit:

ae85abd5aca33c256062cd9048b002525ee7ab8b

We should _never_ be removing the lockfile. This file serves a very
crucial role in ensuring that the site functions in a consistent manner
on different machines. It's possible that this is why our recent pull
requests have stopped working.