Add a bit to the NRQL recommendation for creation
We should mention what the options are, here, and so far we only mention `EVENT_FLOW`. `EVENT_TIMER` is also an option, but nobody would know it from reading this doc.

Should we also mention that `CADENCE` is ALSO an option? We don't think this one works as well as the others, and would like to steer users away from it, but it IS a valid option. Not sure on this one...