docs: add language identifiers to codeblocks
I added java and yml language types to the codeblocks in this section so they can use language specific syntax highlighting if/when that is enabled on our docs. I also added a line of white-space before and after code blocks. Depending on how the md is interpreted the codeblock syntax highlighting can be broken without the white-space before it.