feat: Add k8s metadata injection list of env vars
The list of environment variables that are required for the k8s-apm
linkage is hidden at the bottom of the page as an example of what to
look for. Also, the example is missing the deployment related env var
because it creates a pod directly.

The k8s metadata injection page is:

src/content/docs/kubernetes-pixie/kubernetes-integration/link-your-applications/link-your-applications-kubernetes.mdx