Update Ubuntu release names to match the official
A proposition to update the ubuntu releases to match the ones on the official site for release cycles https://ubuntu.com/about/release-cycle
The reason behind this is because as someone knew to ubuntu and using ubuntu, I am familiar with the official releases. I thought that since it confused me at first, looking at the NR docs, it may confuse others also. 

If there is a particular reason why the names are as they are on the docs please let me know. :) 
Thank you!