docs: reorganize overloads/params table
This PR is a test of a redesign based on a conversation with Robert Siebens, because I'm not able to preview the table changes locally I need to see the Gatsby build to find any mistakes. 

The original table of params was kind difficult to read. I separated the table for each of the 4 overloads (technically could be 6, because the `IDictionary`s could be string/string or string/object but it didn't feel helpful to do that). I also separated `NoticeError(Exception, IDictionary)` and `NoticeError(Exception)`, the IDictionary is optional, but it was harder to read this way, also the `bool $is_expected` param in the later examples is an optional param and we separated that out for readability, so it felt consistent to do it this way. Ultimately the doc is more verbose but I believe it is easier to read.

- added the key/value type options for IDictionary, found in [the source code here](https://github.com/newrelic/newrelic-dotnet-agent/blob/main/src/Agent/NewRelic.Api.Agent/NewRelic.cs)
- made all new headers uniquely hyperlink-able
- Exception is a class, so capitalized it's type. 
- Made indentation 4 spaces for consistency with other C# docs. 
- Changed _Syntax_ section to _Overloads_ to make it more like the .NET MSDocs 
- Added `;` to the Overloads. This is only to help the syntax highlighting on the docs page which comes out a little poor currently.