Remove the availability SLI for browser apps
We've noticed that the suggested queries for the Browser app availability service level are not working well. The reason is that one Page View event can generate more than one Javascript error event, therefore the number of bad events is not always a subset of valid events. Until we come up with a better option, we're removing this section (I'm commenting the content so we can reuse it quickly if needed).