Query would not have worked and is not faceted
Bad query example, since there was no actual FACET clause in the inner query. I edited it and also added quotes around the string value that must have quotes. I also used a real event type (SystemSample) for the example instead of a pretend one, as well as changing the attribute to match one on that event type.