Rename availability as success
The 'availability' term has generated some confusion because it usually refers to a service being up and running and able to receive requests. This is why we're renaming the suggested SLI as Success from now on.

I haven't changed the anchor id, to prevent breaking any existing links.

We'll publish a release note as well.