Update statsd-monitoring-integration-version-2.mdx
To allow integrations with Consul- Hashicorp

The service name registered in Consul will be set to the name of the Kubernetes service associated with the Pod. 
This can be customised with the consul.hashicorp.com/connect-service annotation. 
If using ACLs, this name must be the same as the Pod's ServiceAccount name.

The service account name on the deployment must match the name of the Kubernetes Service or the consul.hashicorp.com/connect-service annotation, default would not match the service name newrelic-statsd.