Update net-agent-configuration.mdx
Added some environment variable configuration subtleties in Windows including that `NEWRELIC_INSTALL_PATH` should set to `C:\Program Files\New Relic\.NET Agent\` which is where the agent DLLs are placed on the system, and `CORECLR_NEWRELIC_HOME` is set to `C:\ProgramData\New Relic\.NET Agent\` which is where configuration and logs exist on the system, and that the Nuget packages use a different set of environment variables. Linux install and Nuget packages use the same set.