Update logs doc with tarball instruction
Many customers are forced to implement infra agents using the tarball installation. Unfortunately, the tarball doesn't contain the td-agent-bit binary (or any other integration), so if a customer still wants to use the log forwarding feature - they are left alone to figure out how to do that. The process is simple enough if you know it, so why not document it :)