Scoped view highlight & OTel support
A few changes on this PR: Scoped view should NOT be at the bottom. We want to call out the option that doesn't require setup right away. The image in the scoped view should be changed because it doesn't actually show scoped view (will submit image on Slack thread). We're going to be providing OTel support (data type) so I updated the main sentence under the scoped views section.