docs: fixed many formatting errors
There were a lot of codeblocks that shouldn't have been there or contained markdown for the page. Most caused by indentation (4 spaces in md makes a codeblock). I also added language identifiers.

I suspect this part is incorrect, it says it's a `.config` file but the code example is c# not xml

     **Instantiating Logger using `.config` file**

     ```cs
     var logger = LogManager.GetLogger("NewRelicLog");
     logger.Info("Hello, New Relic!");
     ```