Fix download links to be HTTPS
While attempting to use the provided download links I got greeted by an error message in Edge (Chromium) that it couldn't be downloaded securely. After figuring out that that meant that the download link was HTTP I figured I might as well create a PR to fix the download links to use https. They seem to work just fine.