Update introduction-log-api.mdx
This change fixes the example which shows "message" being retained if it is JSON. If message is a JSON object, the attributes will be parsed out of it and the original message field will not be retained. Also this fixes the value of service-name to be the same as the example above it.