Added EKS Anywhere on bare metal
This is Kevin Downs, New Relic PMM for infrastructure - kdowns@newrelic.com.

I updated the EKS-Anywhere to EKS Anywhere to match AWS naming. I also added EKS Anywhere on bare metal as we support it.
Can this change be made on June 30th please - that is the date AWS will add the Amazon EKS Anywhere on bare metal service.