docs: update confusing language
This line of code has `name` and `Name` in it
```xml
<tracerFactory name="NewRelic.Agent.Core.Tracer.Factories.BackgroundThreadTracerFactory" metricName="Name">
```
It was difficult to understand which was referred to later when `Name` is referred to, so I made `Name` `TransactionName`. I also removed a dead link and updated the text to make more sense. Then I added language identifiers