docs: fix chars used for double quotes
A lot of this code used the wrong unicode double quotes. `”` is not the same as `"` and  `”` is an invalid char in c#, so that code was broken.  
I also fixed some other code wonkyness like a stray `)` breaking a line of code and some lines mashed together on a single line which were tougher to read.