fix(node agent): update the limit for custom events.
Now the default is 30k, up from 1k.

Memory can be increased, but in my testing, it's a very slight
increase. It depends on the complexity of the custom events and the
rate at which they are being generated.

The server currently only accepts a maximum of 100k custom events per
minute, so that's also a hard limit for the agent.