Update net-agent-configuration.mdx
Modified the copy paste section to be more specific to avoid errors of choosing the directory level too deep.  Condense the steps outlined in the text:When using the Windows .NET agent installer:
Ensure that NEWRELIC_INSTALL_PATH is set to C:\Program Files\New Relic\.NET Agent\. This is where the agent DLLs are placed on the system.
Ensure that CORECLR_NEWRELIC_HOME is set to C:\ProgramData\New Relic\.NET Agent\. This is where configuration and logs exist on the system.

to the section that users can copy & paste.  The .NET agent installer should insert the right environment variables, otherwise that's a bug on our end.