Update metric-normalization-rules.mdx
Hello! I'm a 1st shift TSE on the APM team! There is an order field that is used to determine the precedence of created normalization rules, and has been a source of confusion for customers and support engineers. I've spoken with members of the 3rd shift TSE's who are working on the new MGI UI and they've confirmed that a lower number means higher precedence. I think this should be added to the docs!