AIOps Doc Changes
A few changes have been made to 2 major docs. First, a document that was redirected to "Getting Started with Incident Intelligence" is now it's own page because the redirected page needs to be removed. The other change was "Getting Started with Incident Intelligence" was renamed to "Getting Started with Applied Intelligence". The content on the original doc was put into other parts of AIOps docs where relevant and now a newly updated doc with new content is on the page "Getting Started with Applied Intelligence".

Co-Authored-By: Sammie downing <48104650+homelessbirds@users.noreply.github.com>