Updated fb example
Updated the fb example to show the FILTER block in with the INPUT and OUTPUT blocks in step 4. Also removed the section on associating logs with entities since it's redundant to the example in step 4. We also don't explain to customer why associating logs with entities is  important so removed that verbiage entirely.