new aiops doc
This new doc was requested to be created by Matan and David Neth. The document describes the difference between Alert (classic) and Workflows, as well as provides users a new table describing the difference between what old webhook payload names used to be and what the new corresponding names in the issue payload are.

Co-Authored-By: Sammie downing <48104650+homelessbirds@users.noreply.github.com>