Update anomaly-detection-applied-intelligence.mdx
Line 252 - Insert first image sent in slack - It is an image of selecting the workflow data
Line 254 - Insert second image sent in slack showing destinations
Line 286 - Insert third image sent in slack that show issue feed (If we don't think it goes here, we can leave it out)
For review with @homelessbirds 
I sent over a video in slack that gives an overview of the experience of automatic anomalies. I take the advice of the docs team where to link this.