Create cumulative-metrics.mdx
# Work in Progress

Adding documentation for cumulative sum/histogram support for OTLP. There is also [this doc](https://docs.newrelic.com/docs/more-integrations/open-source-telemetry-integrations/opentelemetry/best-practices/opentelemetry-best-practices-metrics/) that will need some updates (like referring to this doc for cumulative support).

Not sure where the best place for this doc to live is yet, as in the future this will also apply to Prometheus Remote Write delta conversion. Definitely open to suggestions and feedback!