fix: the ESD needs an entityGuid to be returned
In order for the Entity Status Daemon (ESD) to find the associated Synthetic monitor, the ESD needs an entityGuid to be returned as part of the data gathered by the alert condition query. Moving the identifying attribute out of the `filter()` function will accomplish this and allow the monitor's status to turn red or green rather than remain grey in Synthetics.