Update apm-sla-reports.mdx
The Service Levels features are no longer in preview. 

I think it would be cool to add a sentence on the end of this to prompt the user to check it Service Levels, but I don't know what exactly to say. 
Something to the effect of "Service Levels allows you to define the most important performance metrics for your business and view results across Workloads"