Update choose-your-aggregation-method.mdx
Our docs provide a clear, concise example for the particular scenario in which you'd want to use the cadence aggregation method. However, we immediately follow it up with a contradictory statement: "If you're currently using cadence, choose one of the other aggregation methods." This edit softens the statement to note that we only advise against cadence outside of the case mentioned above.