feat: add function to generate EOL date
Created a new function called `getEOLDate()` that will programatically generate an EOL date based on a piece of software's release date and our EOL policy, which is described
[here](https://docs.newrelic.com/docs/licenses/end-of-life/notification-changes-new-relic-saas-features-distributed-software/#impact).

Function accepts two inputs: the first is the release date of the piece of software. The function assumes this input is a string in yyyy-MM-dd format. Second input is an override EOL date. If passed, the function will favor this EOL date over the generated one.

The intent of this function is to allow this repo to programatically generate the various EOL pages it supports, to ease the release process burden on NR developers (updating HTML programatically is hard and humans can forget to do manual steps).