android-sdk-api Add note to recordCustomEvent
Problem: Not all object types are supported for event attributes when recording a custom event.  This leads to an error in the log for each attribute.  This is not described in the documentation.
Solution: Add note linking to setAttribute to help centralize and convey to readers, that there is a limited supported set of types.