wip nav animations
this change adjusts the way the nav is collapsed.
when the nav is open, the page layout is the same.
but when the nav is closed, the page layout is changed in two ways.
the first is that the width of the page container is increased
by an amount equal to the sidebar width, making the container
wider than the viewport.
the new containing div with `overflow: clip;` is to prevent
horizontal scrolling from this.
the second change is that the containg div will transition from
`translate: 0` to `translate: calc(-1 * var(--sidebar-width))`,
effectively hiding the sidebar and shifting the entire containg
div to the left so that the entire width (minus the sidebar)
is visible within the viewport.
the page tools also has a transition to keep it in place
while things are sliding around