Update net-agent-compatibility-requirements-net-core.mdx
The section "Target Framework version" correctly states that we support applications targeting .NET Core 2.0, 2.1, 2.2, 3.0, 3.1, and .NET 5.0, 6.0 and 7.0. The purpose of this PR is to make it clear that we support these versions, but some .NET Core versions require specific agent version ranges to be compatible. For example, a target framework of .NET Core 2.0 is only compatible with agent versions 8.19.353.0 up to 9.9.0 and not our latest 10.x versions.