Update snmp-performance-monitoring.mdx
Added some clarification, added proxy config info for package installations, warned about manual installations.

I attempted to put a copy of the neat button from the top of the page that sends you straight to the guided install into the collapser for the manual install, but the preview does not make it clear if I did it right or if it even works at all.  Please remove it if it cannot be placed in a collaper, thank you.