wip: fix plugin errors, get site serving
remark plugins can't be .mjs files, so this change adds .js files
and uses async imports to get around the ESM issue.
`yarn start` is frequently running out of memory, either trying to build
or after running for a while. giving it 12 GB lets BUILD_FOLDERS=mlops
build pretty reliably, but this is with all but the most important
plugins commented out. the issue now is that it can take nearly 10 minutes
after requesting a page for the page to load in the browser.
that's insane. wtf

this change fixes the MDXContainer body/children issue

this change also adds typehint comments to gatsby-node and gatsby-config.
this gives us a little nicer autocomplete

i also found an O(n^2) loop in gatsby-node.js.
i changed this to use a lookup table instead of using `.find` which
should speed up i18n builds