Update errors-inbox.mdx
The link for the Node Agent incorrectly linked to the Java Agent's docs for setting error group callbacks. There's no specific page for this in the Node space so I changed the link to route to our docs around the API method that is used, which gives details and links to an example app further showing how to use it.