Update whats-new-12-21-NRAI-pp.md
Removing verbiage on terms and conditions as we've changed the user types for certain billing plans that are able to accept legal (pre-release or service-specific) terms on behalf their org. We've also updated the logic that adjusts which terms customers actually need to see, based on what they've already agreed to.
Please also update the existing screenshot (will post new picture in replies).