fix attribute deserialization
deserialization-helpers.js was using an older version of unified and
related packages, which was causing it not to recognize `mdxJsx*`
node types. upgrading in that file would've been difficult, because
it would've had to change to an ESM file, as would handlers.js,
and many more files in the dependency chain.
this is a decent workaround — just define the attribute processor in
deserialize-html.mjs, then pass it as an argument through to the
deserialization-helper functions

to get tests working, i had to add `TRANSLATION_TYPE=human` to the test
command, i wasn't able to get mocking the configuration module working