Add sampling to session traces
Added sampling configuration to the session traces page.

Converted the Get Started section into a steps.

TODO: We may still need to add instructions for viewing their current consumption levels during the 30 day preview before ST bytes become billable on 6/8. That's pending the outcome of how we're instructing customers to do so on 5/2.