use file for release notes commit message
also rewrite the release notes script in Node.
the previous implementation had a bug due to passing commit messages
as a big string to a command line argument, which meant that certain
characters in commits could escape and break the script.
this writes the commit messages to a file and uses that with `gh`