chore(apm/php): update steps to verify apt repo signature
apt-key is deprecated because it adds keys to a trusted shared keyring
which in turns causes apt to use all of those keys when verifying
signatures of packages from any repository.  This compromises trust in
package signing.  The proper way is to specify which key should be used
to verify packages from which repository.