Update manage-synthetics-monitors-rest-api.mdx
Reconfigured the callouts at the top of the page to first identify that this API only functions for legacy runtimes.

And added a second callout which identifies the EOL dates and link regarding the above mentioned legacy runtimes.

Reason for proposing: we still get customers using the REST API because it seems familiar, but they don't get errors from the API highlighting its limitations. It may help deflect support inquiries if we more clearly state that the API should no longer be part of a contemporary workflow.