Remove reference to New Relic One
Hey docs folks -- this is a spot fix, but it needs to be made on the hundreds of spots in the docs that refer to the product as "New Relic One." That's not the name for the product that we use anymore (for at least a year). In addition, there's not a home page, so I removed the reference to home page. Getting to transaction 360 isn't straightforward and it's not part of "All Capabilities" if that's what this was referring to — you need to go to the transaction page for a service and open it from there. 

Here's the spot in the style guide that describes how to refer to New Relic: https://newrelic.atlassian.net/wiki/spaces/PD/pages/3804889254/Product+and+capabilities+capitalization+trademarks+and+usage

Note that there's also a guideline not to say "New Relic's" 

Thanks.